pyproject.toml
[project]
name = "livekit-course"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "livekit-agents[mcp,sarvam,silerio,silero,turn-detector]~=1.3",
    "livekit-api>=1.1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.1",
]
